Added header files
Added header files for PokemonType and PokemonChoice
