+++ Pokemon/PokemonChoice.h
@@ -1,0 +1,10 @@
+#pragma once
+
+// defining an enum class for Pokemon choice
+enum class PokemonChoice
+{
+    CHARMANDER = 1,
+    BULBASAUR,
+    SQUIRTLE,
+    PIKACHU // default choice
+};
