+++ Pokemon/PokemonType.h
@@ -1,0 +1,11 @@
+#pragma once
+
+// defining an enum class for Pokemon type
+enum class PokemonType
+{
+    FIRE,
+    GRASS,
+    WATER,
+    ELECTRIC,
+    NORMAL //Added for default constructor
+};
